Offline Plotting GraphML Script
Plotly GraphML ipynb is supposed to take our GraphML and a node CSV
with atlas labels and convert it into a linked graph with edges.

Currently outputs the Plotly, but no edges show up?

Note changes to the localeq.nodes.csv - I added names,x,y,z to the
header to maximize compatibility for the panda data frame.  In the
future, make sure the CSV outputs that as the header!

Also, it’d be nice if instead of ‘s123’ or something we had the actual
atlas designations.  That’s a future feature we can add in.

WIP - getting the atlas label locations, and getting edges to show up.