Significant updates to analysis3
Fixed centroids plot coloration scheme to (attempt to) match the region
plot current color schema.  Added names to the centroids plot to make
it easier to locate stuff.  It’s mostly there (much closer than it used
to be), although the two plots sometimes still show different regions?
Going to have to keep digging in to this.

Also included current up-to-date html outputs.